Moved remaining in source tests to separate files and fixed some type issues
