+++ app/components/tic-tac-toe-board/types.test.ts
@@ -1,0 +1,84 @@
+import { describe, expect, it } from 'vitest'
+import { FilledCellState, isBoardState } from './types'
+
+describe('board state predicate', () => {
+  const X = FilledCellState.X
+  const O = FilledCellState.O
+  const _ = null
+
+  it('accepts array of cell states', () => {
+    const boardState = [
+      X, _, O,
+      X, _, _,
+      O, _, _,
+    ]
+
+    expect(isBoardState(boardState)).toBe(true)
+  })
+
+  it('accepts nested array of cell states', () => {
+    const boardState = [
+      [
+        X, _, O,
+        X, _, _,
+        O, _, _,
+      ],
+      [
+        _, _, _,
+        _, _, _,
+        _, _, _,
+      ],
+      [
+        _, _, _,
+        _, _, _,
+        _, _, _,
+      ],
+      [
+        _, _, _,
+        _, _, _,
+        _, _, _,
+      ],
+      [
+        X, _, O,
+        X, _, _,
+        O, _, _,
+      ],
+      [
+        _, _, _,
+        _, _, _,
+        _, _, _,
+      ],
+      [
+        _, _, _,
+        _, _, _,
+        _, _, _,
+      ],
+      [
+        _, _, _,
+        _, _, _,
+        _, _, _,
+      ],
+      [
+        _, _, _,
+        _, _, _,
+        _, _, _,
+      ],
+    ]
+
+    expect(isBoardState(boardState)).toBe(true)
+  })
+
+  it('does not accept mixed array of cells', () => {
+    const boardState = [
+      _, _, _,
+      _, _, _,
+      _, _, [
+        X, _, O,
+        X, _, _,
+        O, _, _,
+      ],
+    ]
+
+    expect(isBoardState(boardState)).toBe(false)
+  })
+})
