+++ app/components/tic-tac-toe-board/types.ts
@@ -27,89 +27,3 @@ export const isBoardState = (v: unknown): v is BoardState => Array.isArray(v)
 export const assertIsBoardState: (v: unknown) => asserts v is BoardState = (v) => {
   if (!isBoardState(v)) throw new Error(`Board state is invalid: ${JSON.stringify(v)}`)
 }
-
-if (import.meta.vitest) {
-  const { describe, it, expect } = import.meta.vitest
-
-  describe('board state predicate', () => {
-    const X = FilledCellState.X
-    const O = FilledCellState.O
-    const _ = null
-
-    it('accepts array of cell states', () => {
-      const boardState = [
-        X, _, O,
-        X, _, _,
-        O, _, _,
-      ]
-
-      expect(isBoardState(boardState)).toBe(true)
-    })
-
-    it('accepts nested array of cell states', () => {
-      const boardState = [
-        [
-          X, _, O,
-          X, _, _,
-          O, _, _,
-        ],
-        [
-          _, _, _,
-          _, _, _,
-          _, _, _,
-        ],
-        [
-          _, _, _,
-          _, _, _,
-          _, _, _,
-        ],
-        [
-          _, _, _,
-          _, _, _,
-          _, _, _,
-        ],
-        [
-          X, _, O,
-          X, _, _,
-          O, _, _,
-        ],
-        [
-          _, _, _,
-          _, _, _,
-          _, _, _,
-        ],
-        [
-          _, _, _,
-          _, _, _,
-          _, _, _,
-        ],
-        [
-          _, _, _,
-          _, _, _,
-          _, _, _,
-        ],
-        [
-          _, _, _,
-          _, _, _,
-          _, _, _,
-        ],
-      ]
-
-      expect(isBoardState(boardState)).toBe(true)
-    })
-
-    it('does not accept mixed array of cells', () => {
-      const boardState = [
-        _, _, _,
-        _, _, _,
-        _, _, [
-          X, _, O,
-          X, _, _,
-          O, _, _,
-        ],
-      ]
-
-      expect(isBoardState(boardState)).toBe(false)
-    })
-  })
-}
